Revert "New Immersive Interactive template, legacy one now known as FullPageInteractive (merge 1 of 5)"
