+++ dotcom-rendering/index.d.ts
@@ -43,8 +43,7 @@ type CAPIDesign =
 	| 'InteractiveDesign'
 	| 'PhotoEssayDesign'
 	| 'PrintShopDesign'
-	| 'ObituaryDesign'
-	| 'FullPageInteractiveDesign';
+	| 'ObituaryDesign';
 
 // CAPIDisplay is the display information passed through from CAPI and dictates the displaystyle of the content e.g. Immersive
 // https://github.com/guardian/content-api-scala-client/blob/master/client/src/main/scala/com.gu.contentapi.client/utils/format/Display.scala
