+++ dotcom-rendering/src/web/layouts/ImmersiveLayout.tsx
@@ -4,6 +4,8 @@ import {
 	neutral,
 	brandBackground,
 	brandBorder,
+	labs,
+	border,
 } from '@guardian/src-foundations/palette';
 import { from, until } from '@guardian/src-foundations/mq';
 import { space } from '@guardian/src-foundations';
@@ -15,31 +17,36 @@ import { RightColumn } from '@root/src/web/components/RightColumn';
 import { ArticleContainer } from '@root/src/web/components/ArticleContainer';
 import { ArticleMeta } from '@root/src/web/components/ArticleMeta';
 import { SubMeta } from '@root/src/web/components/SubMeta';
+import { MainMedia } from '@root/src/web/components/MainMedia';
 import { ArticleTitle } from '@root/src/web/components/ArticleTitle';
 import { ArticleHeadline } from '@root/src/web/components/ArticleHeadline';
 import { Standfirst } from '@root/src/web/components/Standfirst';
 import { Footer } from '@root/src/web/components/Footer';
 import { SubNav } from '@root/src/web/components/SubNav/SubNav';
 import { ElementContainer } from '@root/src/web/components/ElementContainer';
+import { Nav } from '@root/src/web/components/Nav/Nav';
 import { MobileStickyContainer, AdSlot } from '@root/src/web/components/AdSlot';
 import { Border } from '@root/src/web/components/Border';
 import { GridItem } from '@root/src/web/components/GridItem';
 import { Caption } from '@root/src/web/components/Caption';
 import { HeadlineByline } from '@root/src/web/components/HeadlineByline';
+import { ContainerLayout } from '@root/src/web/components/ContainerLayout';
 import { Discussion } from '@frontend/web/components/Discussion';
 import { Hide } from '@root/src/web/components/Hide';
+import { LabsHeader } from '@frontend/web/components/LabsHeader';
 import { GuardianLabsLines } from '@frontend/web/components/GuardianLabsLines';
 
 import { buildAdTargeting } from '@root/src/lib/ad-targeting';
+import { getZIndex } from '@frontend/web/lib/getZIndex';
 import { parse } from '@frontend/lib/slot-machine-flags';
 
-import { BannerWrapper } from '@root/src/web/layouts/lib/stickiness';
+import { Stuck, BannerWrapper } from '@root/src/web/layouts/lib/stickiness';
 import {
 	decideLineCount,
 	decideLineEffect,
+	getCurrentPillar,
 } from '@root/src/web/lib/layoutHelpers';
 import { Lines } from '@guardian/source-react-components-development-kitchen';
-import { ImmersiveHeader } from './headers/ImmersiveHeader';
 
 const ImmersiveGrid = ({ children }: { children: React.ReactNode }) => (
 	<div
@@ -164,6 +171,22 @@ const stretchLines = css`
 		margin-right: -10px;
 	}
 `;
+
+const hasMainMediaStyles = css`
+	height: 100vh;
+	/**
+    100vw is normally enough but don't let the content shrink vertically too
+    much just in case
+    */
+	min-height: 25rem;
+	${from.desktop} {
+		min-height: 31.25rem;
+	}
+	${from.wide} {
+		min-height: 50rem;
+	}
+`;
+
 interface Props {
 	CAPI: CAPIType;
 	NAV: NavType;
@@ -184,6 +207,43 @@ const decideCaption = (mainMedia: ImageBlockElement): string => {
 		caption.push(mainMedia.data.credit);
 	return caption.join(' ');
 };
+
+const Box = ({
+	palette,
+	children,
+}: {
+	palette: Palette;
+	children: React.ReactNode;
+}) => (
+	<div
+		css={css`
+			/*
+				This pseudo css shows a black box to the right of the headline
+				so that the black background of the inverted text stretches
+				all the way right. But only from mobileLandscape because below
+				that we want to show a gap. To work properly it needs to wrap
+				the healine so it inherits the correct height based on the length
+				of the headline text
+			*/
+			${from.mobileLandscape} {
+				position: relative;
+				:after {
+					content: '';
+					display: block;
+					position: absolute;
+					width: 50%;
+					right: 0;
+					background-color: ${palette.background.headline};
+					${getZIndex('immersiveBlackBox')}
+					top: 0;
+					bottom: 0;
+				}
+			}
+		`}
+	>
+		{children}
+	</div>
+);
 
 export const ImmersiveLayout = ({
 	CAPI,
@@ -224,14 +284,149 @@ export const ImmersiveLayout = ({
 	const captionText = decideCaption(mainMedia);
 	const { branding } = CAPI.commercialProperties[CAPI.editionId];
 
+	const HEADLINE_OFFSET = mainMedia ? 120 : 0;
+
+	const LeftColCaption = () => (
+		<div
+			css={css`
+				margin-top: ${HEADLINE_OFFSET}px;
+				position: absolute;
+				margin-left: 20px;
+			`}
+		>
+			<Caption
+				palette={palette}
+				captionText={captionText}
+				format={format}
+				shouldLimitWidth={true}
+				isLeftCol={true}
+			/>
+		</div>
+	);
+
 	return (
 		<>
-			<ImmersiveHeader
-				CAPI={CAPI}
-				NAV={NAV}
-				format={format}
-				palette={palette}
-			/>
+			<header
+				css={css`
+					background-color: ${palette.background.article};
+				`}
+			>
+				<div
+					css={[
+						mainMedia && hasMainMediaStyles,
+						css`
+							display: flex;
+							flex-direction: column;
+						`,
+					]}
+				>
+					<div
+						css={css`
+							${getZIndex('headerWrapper')}
+							order: 0;
+						`}
+					>
+						<ElementContainer
+							showSideBorders={false}
+							showTopBorder={false}
+							padded={false}
+							backgroundColour={brandBackground.primary}
+							element="nav"
+						>
+							<Nav
+								format={{
+									...format,
+									theme: getCurrentPillar(CAPI),
+								}}
+								nav={NAV}
+								subscribeUrl={
+									CAPI.nav.readerRevenueLinks.header.subscribe
+								}
+								edition={CAPI.editionId}
+							/>
+						</ElementContainer>
+					</div>
+
+					{format.theme === ArticleSpecial.Labs && (
+						<Stuck>
+							<ElementContainer
+								showSideBorders={true}
+								showTopBorder={false}
+								backgroundColour={labs[400]}
+								borderColour={border.primary}
+								sectionId="labs-header"
+							>
+								<LabsHeader />
+							</ElementContainer>
+						</Stuck>
+					)}
+
+					<MainMedia
+						format={format}
+						palette={palette}
+						elements={CAPI.mainMediaElements}
+						adTargeting={adTargeting}
+						starRating={
+							format.design === ArticleDesign.Review &&
+							CAPI.starRating
+								? CAPI.starRating
+								: undefined
+						}
+						host={host}
+						hideCaption={true}
+						pageId={CAPI.pageId}
+						webTitle={CAPI.webTitle}
+					/>
+				</div>
+				{mainMedia && (
+					<>
+						<div
+							css={css`
+								margin-top: -${HEADLINE_OFFSET}px;
+								/*
+                        This z-index is what ensures the headline title text shows above main media. For
+                        the actual headline we set the z-index deeper in ArticleHeadline itself so that
+                        the text appears above the pseudo Box element
+                    */
+								position: relative;
+								${getZIndex('articleHeadline')};
+							`}
+						>
+							<ContainerLayout
+								verticalMargins={false}
+								padContent={false}
+								padSides={false}
+								leftContent={<LeftColCaption />}
+							>
+								<ArticleTitle
+									format={format}
+									palette={palette}
+									tags={CAPI.tags}
+									sectionLabel={CAPI.sectionLabel}
+									sectionUrl={CAPI.sectionUrl}
+									guardianBaseURL={CAPI.guardianBaseURL}
+									badge={CAPI.badge}
+								/>
+							</ContainerLayout>
+							<Box palette={palette}>
+								<ContainerLayout
+									verticalMargins={false}
+									padContent={false}
+									padSides={false}
+								>
+									<ArticleHeadline
+										format={format}
+										headlineString={CAPI.headline}
+										palette={palette}
+										tags={CAPI.tags}
+										byline={CAPI.author.byline}
+									/>
+								</ContainerLayout>
+							</Box>
+						</div>
+					</>
+				)}
+			</header>
 			<main>
 				<ElementContainer
 					showTopBorder={false}
