--- dotcom-rendering/src/web/layouts/InteractiveImmersive.stories.tsx
@@ -1,58 +1,0 @@
-import { useEffect } from 'react';
-
-import {
-	makeGuardianBrowserCAPI,
-	makeGuardianBrowserNav,
-} from '@root/src/model/window-guardian';
-import { Article } from '@root/fixtures/generated/articles/Article';
-
-import { BootReact } from '@root/src/web/components/BootReact';
-import { embedIframe } from '@root/src/web/browser/embedIframe/embedIframe';
-import { mockRESTCalls } from '@root/src/web/lib/mockRESTCalls';
-
-import { extractNAV } from '@root/src/model/extract-nav';
-import { fireAndResetHydrationState } from '@root/src/web/components/HydrateOnce';
-import { DecideLayout } from './DecideLayout';
-
-mockRESTCalls();
-
-export default {
-	title: 'Layouts/InteractiveImmersive',
-	parameters: {
-		chromatic: { viewports: [1300], delay: 800, diffThreshold: 0.2 },
-	},
-};
-
-const convertToInteractiveImmersive = (CAPI: CAPIType) => {
-	return {
-		...CAPI,
-		format: {
-			...CAPI.format,
-			display: 'ImmersiveDisplay' as CAPIDisplay,
-			design: 'InteractiveDesign' as CAPIDesign,
-		},
-	};
-};
-
-// HydratedLayout is used here to simulated the hydration that happens after we init react on
-// the client. We need a separate component so that we can make use of useEffect to ensure
-// the hydrate step only runs once the dom has been rendered.
-const HydratedLayout = ({ ServerCAPI }: { ServerCAPI: CAPIType }) => {
-	fireAndResetHydrationState();
-	const NAV = extractNAV(ServerCAPI.nav);
-
-	useEffect(() => {
-		const CAPI = makeGuardianBrowserCAPI(ServerCAPI);
-		BootReact({ CAPI, NAV: makeGuardianBrowserNav(NAV) });
-		embedIframe().catch((e) =>
-			console.error(`HydratedLayout embedIframe - error: ${e}`),
-		);
-	}, [ServerCAPI, NAV]);
-	return <DecideLayout CAPI={ServerCAPI} NAV={NAV} />;
-};
-
-export const ArticleStory = (): React.ReactNode => {
-	const ServerCAPI = convertToInteractiveImmersive(Article);
-	return <HydratedLayout ServerCAPI={ServerCAPI} />;
-};
-ArticleStory.story = { name: 'Article' };
